Introduce type-safe units from the [squants] library
Currently, lengths, angles and time are all just Doubles, which introduces
a possibility for error. Scala's type system allows for very rich units
which the [squants] library leverages.

Although STL, OpenSCAD and others are "unit-less" where it comes to
lengths, they're not unit-less on angles. Plus, it makes sense to be
specific where possible, e.g. when mixing inches and mm in one project.

This commit introduces type-safe units in most places, but leaves the
examples as-is, interpreting everything in millimeter or radians.